pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.0.4",
    "fastapi>=0.115.13",
    "googlesearch-python>=1.3.0",
    "groq>=0.29.0",
    "openai>=1.91.0",
    "packaging>=25.0",
    "phidata>=2.7.10",
    "pycountry>=24.6.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
    "uvicorn>=0.34.3",
    "yfinance>=0.2.63",
]
